+++ Source/ChilliSource/Rendering/Particle/Drawable/StaticBillboardParticleDrawable.cpp
@@ -172,7 +172,7 @@ namespace ChilliSource
 				}
 				break;
 			case StaticBillboardParticleDrawableDef::ImageSelectionType::k_random:
-				m_particleBillboardIndices[in_index] = Core::Random::Generate<u32>(0, m_billboards->size() - 1);
+				m_particleBillboardIndices[in_index] = Core::Random::Generate<u32>(0, static_cast<u32>(m_billboards->size()) - 1);
 				break;
 			default:
 				CS_LOG_FATAL("Invalid image selection type.");
